+++ src/test/java/wooteco/subway/acceptance/StationAcceptanceTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.transaction.annotation.Transactional;
-import wooteco.subway.acceptance.step.StationRequest;
+import wooteco.subway.acceptance.request.StationRequest;
 
 import java.util.Map;
